// src/main/resources/fleet-core-frontend/src/app/components-2/admin/admin-dashboard/models/vehicle.model.ts
export class Vehicle{
        id: number;
        make: string;
        model: string;
        vin: string;
        lincesePlate: string;
        milenage: number;
        fuelType: string;     
        yearOfManufacture: number;          
        cargoCapacity: number;  
        fuelCapacity: number;
        fuelConsumption: number;     
        vehicleStatus: string;
        imageData: string
        adminId: number
}